+++ code/go/0chain.net/smartcontract/zcnsc/nodes_reader.go
@@ -5,26 +5,10 @@ import (
 	"reflect"
 
 	. "0chain.net/chaincore/config"
-	"0chain.net/chaincore/state"
 	"0chain.net/core/util"
 
 	cstate "0chain.net/chaincore/chain/state"
 	"0chain.net/core/common"
-)
-
-const (
-	SmartContract = "smart_contracts"
-	ZcnSc         = "zcnsc"
-)
-
-const (
-	MinMintAccount     = "min_mint_account"
-	PercentAuthorizers = "percent_authorizers"
-	MinAuthorizers     = "min_authorizers"
-	MinBurnAmount      = "min_burn_amount"
-	MinStakeAmount     = "min_stake_amount"
-	BurnAddress        = "burn_address"
-	MaxFee             = "max_fee"
 )
 
 var (
@@ -99,17 +83,7 @@ func GetGlobalNode(ctx cstate.StateContextI) (*GlobalNode, error) {
 		return nil, err
 	}
 
-	gn.MinMintAmount = state.Balance(cfg.GetInt(Section(MinMintAccount)))
-	gn.PercentAuthorizers = cfg.GetFloat64(Section(PercentAuthorizers))
-	gn.MinAuthorizers = cfg.GetInt64(Section(MinAuthorizers))
-	gn.MinBurnAmount = cfg.GetInt64(Section(MinBurnAmount))
-	gn.MinStakeAmount = cfg.GetInt64(Section(MinStakeAmount))
-	gn.BurnAddress = cfg.GetString(Section(BurnAddress))
-	gn.MaxFee = cfg.GetInt64(Section(MaxFee))
+	gn.Config = loadSettings()
 
 	return gn, nil
-}
-
-func Section(section string) string {
-	return fmt.Sprintf("%s.%s.%s", SmartContract, ZcnSc, section)
 }
